sqlserver/sqlserver.go
package sqlserver

import (
	"errors"
	"strings"

	"github.com/mstgnz/sqlporter"
)

type SQLServer struct {
	schema *sqlporter.Schema
}

// NewSQLServer creates a new SQLServer parser instance
func NewSQLServer() *SQLServer {
	return &SQLServer{
		schema: &sqlporter.Schema{},
	}
}

// Parse parses SQLServer dump content
func (s *SQLServer) Parse(content string) (*sqlporter.Schema, error) {
	if content == "" {
		return nil, errors.New("empty content")
	}

	// TODO: SQLServer dump parsing operations will be implemented here
	return s.schema, nil
}

// Generate generates SQLServer dump from schema
func (s *SQLServer) Generate(schema *sqlporter.Schema) (string, error) {
	if schema == nil {
		return "", errors.New("empty schema")
	}

	var result strings.Builder

	// TODO: SQLServer dump generation operations will be implemented here

	return result.String(), nil
}
